# docs/swagger.yaml
basePath: /
definitions:
...
  httpx.ErrorDetail:
    properties:
      field:
        type: string
      message:
        type: string
    type: object
  httpx.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/httpx.ErrorResponseBody'
      success:
        type: boolean
    type: object
  httpx.ErrorResponseBody:
    properties:
      code:
        type: string
      details:
        items:
          $ref: '#/definitions/httpx.ErrorDetail'
        type: array
      message:
        type: string
    type: object
  httpx.SuccessResponse:
    properties:
      data: {}
      meta: {}
      success:
        type: boolean
    type: object
  profile.UpdateCommand:
    properties:
      bio:
        type: string
      is_public:
        type: boolean
      location:
        type: string
      reading_preferences:
        items:
          type: integer
        type: array
      username:
        type: string
      website:
        type: string
    type: object
  rating.createRatingReq:
    properties:
      star:
        maximum: 5
        minimum: 1
        type: integer
    required:
    - star
    type: object
  readinglist.upsertReq:
    properties:
      isbn:
        type: string
      status:
        type: string
    required:
    - isbn
    - status
    type: object
...
host: localhost:8080
info:
  contact:
    name: API Support
  description: A REST API for tracking books you've read, want to read, or are currently
    reading.
  termsOfService: http://swagger.io/terms/
  title: Personal Book Tracking API
  version: "1.0"
paths:
...
  /books/{isbn}/rating:
    get:
      consumes:
      - application/json
      description: Get average rating and total count for a book
      parameters:
      - description: Book ISBN
        in: path
        name: isbn
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpx.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      summary: Get book rating
      tags:
      - ratings
    post:
      consumes:
      - application/json
      description: Rate a book (1-5 stars)
      parameters:
      - description: Book ISBN
        in: path
        name: isbn
        required: true
        type: string
      - description: Rating request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/rating.createRatingReq'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      security:
      - Bearer: []
      summary: Create or update book rating
      tags:
      - ratings
...
  /me/profile:
    get:
      consumes:
      - application/json
      description: Get the authenticated user's complete profile
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpx.SuccessResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      security:
      - Bearer: []
      summary: Get own profile
      tags:
      - profiles
    patch:
      consumes:
      - application/json
      description: Update the authenticated user's profile
      parameters:
      - description: Profile update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/profile.UpdateCommand'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpx.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      security:
      - Bearer: []
      summary: Update own profile
      tags:
      - profiles
  /me/sessions:
    get:
      consumes:
      - application/json
      description: Get all active sessions for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpx.SuccessResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      security:
      - Bearer: []
      summary: List user sessions
      tags:
      - sessions
  /me/sessions/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a specific session for the authenticated user
      parameters:
      - description: Session ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      security:
      - Bearer: []
      summary: Delete session
      tags:
      - sessions
...
  /users/{id}/{status}:
    get:
      consumes:
      - application/json
      description: Get a user's reading list filtered by status (WISHLIST, READING,
        FINISHED)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Reading list status
        enum:
        - WISHLIST
        - READING
        - FINISHED
        in: path
        name: status
        required: true
        type: string
      - default: 20
        description: Limit results
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset for pagination
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpx.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      summary: List books by status
      tags:
      - reading-lists
  /users/{id}/profile:
    get:
      consumes:
      - application/json
      description: Get a user's public profile information
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpx.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      summary: Get public profile
      tags:
      - profiles
...
  /users/readinglist:
    post:
      consumes:
      - application/json
      description: Add a book to reading list or update its status (WISHLIST, READING,
        FINISHED)
      parameters:
      - description: Reading list request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/readinglist.upsertReq'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      security:
      - Bearer: []
      summary: Add or update reading list entry
      tags:
      - reading-lists
...
schemes:
- http
- https
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"